moved api logic outside sudoku service
- also restructured src/app/core/services
- some reformatting
