+++ frontend/src/app/core/services/sudoku-api/sudoku-api.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SudokuApiService } from './sudoku-api.service';
+
+describe('SudokuApiService', () => {
+  let service: SudokuApiService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SudokuApiService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
