+++ frontend/src/app/core/services/sudoku/sudoku.service.ts
@@ -1,10 +1,7 @@
-import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Cage } from 'src/app/models/cage.model';
 import { CellContent } from 'src/app/models/cell.model';
 import { Coordinates } from 'src/app/models/coordinates.model';
-import { environment } from 'src/environments/environment';
-import { Observable } from 'rxjs';
 
 @Injectable({
 	providedIn: 'root',
@@ -15,8 +12,6 @@ export class SudokuService {
 	public highlightedCells: Coordinates[] = [];
 	private inputDisabled: boolean = false;
 	public inputMethod: 'prefill' | 'solve' = 'solve';
-
-	private basePath = `${environment.protocol}://${environment.backend_url}`;
 
 	possibleSums = [
 		[1, 9],
@@ -30,7 +25,7 @@ export class SudokuService {
 		[45, 45]
 	];
 
-	constructor(private httpClient: HttpClient) {
+	constructor() {
 		this.resetState();
 	}
 
@@ -217,28 +212,5 @@ export class SudokuService {
 		});
 
 		this.grid[highestCell][leftmostHighest].cage.sum = cage.sum;
-	}
-
-	sendAsStartingGrid(): Observable<string> {
-		const startingGrid: number[][] = []
-		this.grid.forEach(row => {
-			startingGrid.push(
-				row.map(cell => cell.values.length > 0 ? cell.values[0] : 0)
-			);
-		});
-		this.cages.forEach(cage => {
-			cage.cells.forEach(cell => {
-				delete cell['adjacency']
-			})
-		})
-
-		return this.httpClient.post(
-			`${this.basePath}/games`,
-			{
-				grid: startingGrid,
-				cages: this.cages
-			},
-			{ responseType: 'text' }
-		);
 	}
 }
